geo_sim_util.py addition
Put in the file geo_sim_util.py  classes and functions that are used by the new version of simplify.py
